# openapi.yaml
openapi: 3.1.0
info:
  title: Persistio API
  version: 0.1.19
  description: Self-hosted memory service for AI agents.
servers:
  - url: http://localhost:4827
paths:
...
  /v1/extract:
    post:
      summary: Trigger extraction
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Extraction job queued
  /v1/jobs/{id}:
    get:
      summary: Check extraction job status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
        '404':
          description: Job not found
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key